+++ docs/changelog.md
@@ -2,6 +2,23 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.33.0] -- 2022-07-25
+
+### Changed
+
+- `pep_version` is no longer a required parameter to create a `peppy.Project` instance from a configuration file.
+
+### Fixed
+
+- Performance issues during sample parsing. Two list comprehensions were combined to speed up this functionality.
+- `KeyError` is thrown when attempting to access the `pep_version` of a `peppy.Project` instance instatiated from a sample table (`csv`)
+
+### Added
+
+- Implementation of `__eq__` for the `peppy.Project` class such that two instances of the class can be compared using python's equality operators (`==`, `!=`).
+- New `from_dict` function that lets a user instatiate a new `peppy.Project` object using an in-memory representation of a PEP (a `dict`). This supports database storage of PEPs.
+- New `extended` flag for the `to_dict` method on `peppy.Project` objects. This creates a **richer** dictionary representation of PEPs.
+- Better sample parsing
 
 ## [0.32.0] -- 2022-05-03